pyproject.toml
[tool.poetry]
name = "veenrokdalv"
version = "0.1.0"
description = ""
authors = ["veenrokroot <veenrokdalv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
python-environ = "^0.4.54"
drf-spectacular = "^0.24.2"
drf-spectacular-sidecar = "^2022.10.1"
djangorestframework-camel-case = "^1.3.0"
psycopg2-binary = "^2.9.4"
django-bootstrap-v5 = "^1.0.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
